// Live/frontend/src/client/models/UserOut.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type UserOut = {
    email: string;
    is_active?: boolean;
    is_superuser: "admin" | "user" | "sub-user";
    first_name?: (string | null);
    id: number;
};